requirements.txt
alembic==1.7.3
autopep8==1.5.7
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.1
zipp==3.5.0
